pyproject.toml
[project]
name = "voiceslice"
version = "0.1.0"
description = "音频切片和文本识别工具"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "VoiceSlice Contributors"}
]

dependencies = [
    # 音频处理
    "numpy<2.0",
    "scipy",
    "librosa==0.10.2",
    "soundfile",
    "ffmpeg-python",
    
    # ASR
    "faster-whisper>=1.1.1",
    "funasr==1.0.27",
    "torch",
    "torchaudio",
    
    # WebUI
    "gradio<5",
    
    # 工具
    "tqdm",
    "huggingface-hub>=0.13,<0.20",
    "typer",
    "pyyaml",
]

[project.optional-dependencies]
gpu = [
    "torch",
    "torchaudio",
]

[project.scripts]
voiceslice = "webui.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src", "webui"]

[tool.uv.sources]
# 当使用 --extra gpu 时，从 PyTorch CUDA 索引安装 GPU 版本
# 注意：这会影响所有安装，包括默认安装
# 如果需要 CPU 版本，请使用 uv sync（不使用 --extra gpu）
torch = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
    { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
